pyproject.toml
[project]
name = "styleclaw"
version = "0.1.0"
description = "AI-powered style trigger word exploration system"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "boto3>=1.35",
    "pydantic>=2.9",
    "typer>=0.15",
    "pillow>=11.0",
    "python-dotenv>=1.0",
    "jinja2>=3.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "respx>=0.22",
    "coverage>=7.0",
]

[project.scripts]
styleclaw = "styleclaw.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/styleclaw"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: real API calls (deselect with '-m \"not integration\"')",
]

[tool.coverage.run]
source = ["src/styleclaw"]

[tool.coverage.report]
fail_under = 80
